+++ package.json
@@ -14,6 +14,7 @@
     "@dnd-kit/utilities": "^3.2.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-icons": "^5.5.0",
     "react-router-dom": "^7.2.0"
   },
   "devDependencies": {
