+++ src/Hero/Hero.css
@@ -1,0 +1,154 @@
+.hero-container {
+    display: flex;
+    height: 100vh;
+    
+  }
+  
+  .sidebar {
+    background-color: #f4f4f4;
+    display: flex;
+    flex-direction: column;
+    justify-content: center; /* Center buttons vertically */
+    align-items: center;
+    width: 600px;
+    gap: 30px;
+    position: relative; /* For dropdown positioning */
+  }
+  
+  .sidebar-button {
+    width: 320px;
+    background: linear-gradient(135deg, #6e8efb, #a777e3);
+    color: white;
+    font-size: 16px;
+    font-weight: bold;
+    padding: 12px 24px;
+    border: none;
+    border-radius: 8px;
+    cursor: pointer;
+    transition: all 0.3s ease-in-out;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
+    text-transform: uppercase;
+  }
+  
+  .sidebar-button:hover {
+    background: linear-gradient(135deg, #5a75f0, #9461d6);
+    transform: translateY(-2px);
+    box-shadow: 0 6px 15px rgba(0, 0, 0, 0.25);
+  }
+  
+  .sidebar-button:active {
+    transform: translateY(1px);
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
+  }
+  
+  .content {
+    display: flex;
+    width: 100%;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding: 0px 120px 0 100px;
+    background: linear-gradient(135deg, #00b8b8, #787878);
+  }
+  
+  .content h1 {
+    font-size: 48px;
+    font-weight: bold;
+    color: #101a24;
+    margin-bottom: 20px;
+    text-transform: uppercase;
+    letter-spacing: 2px;
+    text-shadow: 2px 2px 10px rgba(0, 0, 0, 0.2);
+  }
+  
+  .content p {
+    font-size: 18px;
+    color: #34495e;
+    line-height: 1.8;
+    text-shadow: 1px 1px 5px rgba(0, 0, 0, 0.1);
+  }
+  
+  .content-text {
+    background-color: rgba(240, 248, 255, 0.192);
+    padding: 20px;
+    border-radius: 20px;
+  }
+  
+  .user-info {
+    display: flex;
+    align-items: center;
+    background: #fff;
+    padding: 10px 15px;
+    gap: 170px;
+    border-radius: 50px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+    font-family: Arial, sans-serif;
+    cursor: pointer;
+    position: absolute; /* Position user-info absolutely */
+    top: 30px; /* Adjust top position */
+    left: 50%; /* Center horizontally */
+    transform: translateX(-50%); /* Center horizontally */
+  }
+  
+  .user-icon {
+    width: 55px;
+    height: 40px;
+    border-radius: 50%;
+    background: #a777e3;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: white;
+    font-size: 20px;
+    margin-right: 10px;
+  }
+  
+  .user-name {
+    font-size: 16px;
+    font-weight: 600;
+    color: #333;
+  }
+  
+  /* Dropdown Menu Styles */
+  .dropdown-menu {
+    position: absolute;
+    top: 100px; /* Position below user-info */
+    left: 27%; /* Center horizontally */
+    
+    background: #fff;
+    border-radius: 10px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+    width: 200px;
+    overflow: hidden;
+    animation: slideDown 0.3s ease-out;
+    z-index: 10; /* Ensure it appears above other elements */
+  }
+  
+  .dropdown-item {
+    display: flex;
+    align-items: center;
+    padding: 10px 15px;
+    color: #333;
+    cursor: pointer;
+    transition: background 0.2s;
+  }
+  
+  .dropdown-item:hover {
+    background: #f1f1f1;
+  }
+  
+  .dropdown-icon {
+    margin-right: 10px;
+  }
+  
+  /* Animation for dropdown */
+  @keyframes slideDown {
+    from {
+      opacity: 0;
+      transform: translateY(-20px);
+    }
+    to {
+      opacity: 1;
+      transform: translateY(0);
+    }
+  }
