+++ src/main.jsx
@@ -2,9 +2,9 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
-
+import Hero from './Hero/Hero.jsx'
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <Hero />
   </StrictMode>,
 )
