+++ src/pages/Device.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function DeviceContent() {
+  return (
+    <div>
+      <h1>Device Content</h1>
+      <p>This is the content for the Device section.</p>
+    </div>
+  );
+}
+
+export default DeviceContent;
