+++ src/pages/SlidesContent.jsx
@@ -1,0 +1,13 @@
+// SlidesContent.js
+import React from 'react';
+
+function SlidesContent() {
+  return (
+    <div>
+      <h1>Slides Content</h1>
+      <p>This is the content for the Slides section.</p>
+    </div>
+  );
+}
+
+export default SlidesContent;
