+++ src/pages/TemplatesContent.jsx
@@ -1,0 +1,13 @@
+// TemplatesContent.js
+import React from 'react';
+
+function TemplatesContent() {
+  return (
+    <div>
+      <h1>Templates Content</h1>
+      <p>This is the content for the Templates section.</p>
+    </div>
+  );
+}
+
+export default TemplatesContent;
